% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rmd.R
\name{count_rmds}
\alias{count_rmds}
\title{Check for duplicates in \code{_bookdown.yml} and any left out \code{Rmds}}
\usage{
count_rmds()
}
\value{
duplicates in \code{_bookdown.yml} and any left out \code{Rmds}
}
\description{
Check for duplicates in \code{_bookdown.yml} and any left out \code{Rmds}
}
\examples{
count_rmds()

}
