% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtitle.R
\name{fixtitle}
\alias{fixtitle}
\title{Easily edit bookdown chapter tiles}
\usage{
fixtitle(find, replace = NULL, write = FALSE)
}
\arguments{
\item{find}{full or partial title string to search for}

\item{replace}{replacement string (optional)}

\item{write}{logical indicating whether to change the title,
defaults to FALSE}
}
\description{
In a large project project particularly with multiple authors,
the simple task of editing a chapter title can be difficult since
a few steps are required: find the right Rmd, open the file,
change the title, save it, close it. The \code{fixtitle()} function
simplifies the process by doing all this in one line of code. It can
be used for searching only, or searching and replacing.
}
\examples{
\dontrun{
fixtitle(find = "Instructions")    # search only

fixtitle(find = "Instructions",    # search and replace test
         replace = "Submission instructions")

fixtitle(find = "Instructions",    # search and replace actual
         replace = "Submission instructions",
         write = TRUE)
}


}
