% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pairs.R
\name{make_pairs}
\alias{make_pairs}
\title{Make problem set pairings based on section}
\usage{
make_pairs(file = "~/Downloads/PSet 3.csv", prefix = "PSet 3")
}
\arguments{
\item{file}{path to file}

\item{prefix}{prefix for group names}
}
\value{
data frame with new pairs column
}
\description{
file must be a file downloaded from CourseWorks people section with `group_name` column
}
\examples{

df <- make_pairs("~/Downloads/PSet\ 3.csv")

}
