% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movepart.R
\name{movepart}
\alias{movepart}
\title{Move bookdown (PART) to another chapter}
\usage{
movepart(oldfile, newfile)
}
\arguments{
\item{oldfile}{file that currently contains (PART) info}

\item{newfile}{file to get (PART) info}
}
\description{
Move bookdown (PART) to another chapter
}
\examples{
\dontrun{
movepart(oldfile = "Chapter3.Rmd", newfile = "Chapter4.Rmd")
}

}
