% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_month_year.R
\name{random_month_year}
\alias{random_month_year}
\title{Generate a Random Month and Year}
\usage{
random_month_year()
}
\value{
A character string in the format "Month YYYY", e.g., "October 1993".
}
\description{
Returns a randomly selected month and year between February 1967 and the current month.
}
\examples{
random_month_year()

}
