% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_time.R
\name{record}
\alias{record}
\title{Record what you're doing now to stay on track}
\usage{
record()
}
\value{
No return value. Writes to a file as a side effect.
}
\description{
Record what you're doing now to stay on track
}
\examples{
\dontrun{
record()
}

}
